pyproject.toml
[project]
name = "pokeagent-speedrun-switch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.4.4",
    "openai>=2.32.0",
    "opencv-python>=4.13.0.92",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
pokeagent-speedrun-switch = "pokeagent_speedrun_switch.main:main"
